Clean up tests, add variant groups to query
