+++ tests/data/transform/therapeutic/civic_harvester.json
@@ -2160,6 +2160,7 @@
                 }
             ],
             "aliases": [
+                "ERRP",
                 "EGFR",
                 "mENA",
                 "PIG61",
